fix(logging): stop dropping all ERROR events from Sentry
`EventMessageContains` split its decision across two `Filter` methods: a
level probe in `enabled`, and the message match in `event_enabled`, which
is the only one that receives the event's fields. Composing it with
`FilterExt::not` inverted the wrong half. `Not::enabled` negates the level
probe, so every event at the filter's level was rejected before its
message was ever inspected, and `Not::event_enabled` returns `true`
without delegating, leaving the match unreachable.

The four WinTun suppressions therefore dropped every `ERROR` event from
the Sentry layer in every binary, while never suppressing the WinTun
messages they target. Panics were unaffected because they reach Sentry
through the panic integration rather than the tracing layer.

Fold the negation into the filter so its whole decision lives in
`event_enabled`, where the fields are, and name it for what it does.

Related: #9469
